requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-decouple
Pillow
gunicorn
